+++ scripts/macros/2024/actions/genericActions.js
@@ -6,7 +6,10 @@ async function use({trigger, workflow}) {
     let pack = game.packs.get(packId);
     if (!pack) return;
     let index = await pack.getIndex();
-    let selection = await dialogUtils.selectDocumentDialog(workflow.item.name, 'CHRISPREMADES.Macros.GenericActions.Select', index.contents, {sortAlphabetical: true});
+    let documents = index.contents;
+    let circleCast = genericUtils.getCPRSetting('circleCast');
+    if (!circleCast) documents = documents.filter(i => i.name != 'Circle Cast');
+    let selection = await dialogUtils.selectDocumentDialog(workflow.item.name, 'CHRISPREMADES.Macros.GenericActions.Select', documents, {sortAlphabetical: true});
     if (!selection) return;
     let documentData = genericUtils.duplicate(selection.toObject());
     documentData.system.description.value = itemUtils.getItemDescription(documentData.name);